ml_models/requirements.txt
# Core ML libraries
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.21.0
datasets>=2.4.0

# OCR libraries
easyocr>=1.6.2
paddlepaddle>=2.4.0
paddleocr>=2.6.0
pytesseract>=0.3.10

# Computer Vision
opencv-python>=4.6.0
Pillow>=9.2.0
scikit-image>=0.19.0

# Data processing
pandas>=1.4.0
numpy>=1.21.0
scipy>=1.9.0

# ML algorithms
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Utilities
tqdm>=4.64.0
requests>=2.28.0
python-dotenv>=0.20.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.15.0